+++ src/core/model/object.hpp
@@ -96,6 +96,7 @@ public:
 signals:
     void property_changed(const model::BaseProperty* prop, const QVariant& value);
     void removed_from_list();
+    void added_to_list();
 
 protected:
     virtual void on_property_changed(const BaseProperty* prop, const QVariant& value)